Adds Something that is missing
Adds Futil.strictcolorize (that filters Banned colors on smite and gtfo)

Adds Actual Server-TFM features
Adds Wipewarps
adds wipecoreprotectdata
adds spectator
adds rainbownick
adds rainbowtag

+++ src/main/java/me/totalfreedom/totalfreedommod/command/Command_wipeuserdata.java
@@ -1,8 +1,12 @@
 package me.totalfreedom.totalfreedommod.command;
 
 import java.io.File;
+import java.util.List;
 import me.totalfreedom.totalfreedommod.rank.Rank;
 import me.totalfreedom.totalfreedommod.util.FUtil;
+import org.bukkit.Bukkit;
+import static org.bukkit.Bukkit.getServer;
+import org.bukkit.World;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
@@ -21,11 +25,24 @@ public class Command_wipeuserdata extends FreedomCommand
             return true;
         }
 
-        FUtil.adminAction(sender.getName(), "Wiping Essentials playerdata", true);
+        FUtil.adminAction(sender.getName(), "Wiping Essentials and Worlds playerdata", true);
 
-        FUtil.deleteFolder(new File(server.getPluginManager().getPlugin("Essentials").getDataFolder(), "userdata"));
+        for (World w : Bukkit.getWorlds())
+        {
+            if (w.getName().equals(plugin.wm.adminworld.getWorld()))
+            {
+                return true;
+            }
 
-        msg("All playerdata deleted.");
-        return true;
+            FUtil.deleteFolder(new File(server.getPluginManager().getPlugin("Essentials").getDataFolder(), "userdata"));
+            FUtil.deleteFolder(new File(Bukkit.getServer().getWorld(plugin.wm.flatlands.getName()).getWorldFolder().getName() + "playerdata"));
+            FUtil.deleteFolder(new File(Bukkit.getServer().getWorld(plugin.wm.flatlands.getName()).getWorldFolder().getName() + "stats"));
+            FUtil.deleteFolder(new File(Bukkit.getServer().getWorld(w.getName()).getWorldFolder().getName() + "stats"));
+            FUtil.deleteFolder(new File(Bukkit.getServer().getWorld(w.getName()).getWorldFolder().getName() + "playerdata"));
+
+            msg("All playerdata deleted.");
+            return true;
+        }
+        return false;
     }
 }
